# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Tantakatu API
  description: This is the first attempt for create API for Tantakatu
  contact:
    name: Gabriel Morales
    email: gabs2323@outlook.com
  license:
    name: Gabriel Morales
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/Erraticturtle35/Tantakatu/1.0.0
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /currencies:
    post:
      tags:
      - currency
      summary: Create currency
      operationId: create_currency
      requestBody:
        description: Created currency object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Currency'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.currency_controller
  /currencies/{currency_id}:
    get:
      tags:
      - currency
      summary: Get currency by id
      operationId: get_currency_by_id
      parameters:
      - name: currency_id
        in: path
        description: The currency_id thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        400:
          description: Invalid currency supplied
        404:
          description: Currency not found
      x-openapi-router-controller: swagger_server.controllers.currency_controller
    put:
      tags:
      - currency
      summary: Updates currency
      description: This is only can be done by the logged in user
      operationId: update_currency
      parameters:
      - name: currency_id
        in: path
        description: currency_id that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update currency object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Currency'
        required: true
      responses:
        400:
          description: Invalid currency supplied
        404:
          description: Currency not found
      x-openapi-router-controller: swagger_server.controllers.currency_controller
    delete:
      tags:
      - currency
      summary: Delete currency
      description: This can only be done by the logged in user.
      operationId: delete_currency
      parameters:
      - name: currency_id
        in: path
        description: The currency_id that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        400:
          description: Invalid currency supplied
        404:
          description: Currency not found
      x-openapi-router-controller: swagger_server.controllers.currency_controller
  /currencies/{currency_id}/rate:
    post:
      tags:
      - currency
      summary: Create currency rate
      operationId: create_currency_rate
      parameters:
      - name: currency_id
        in: path
        description: The currency_id thats needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Created a currency object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrencyRate'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.currency_controller
  /currencies/rate/{rate_id}:
    get:
      tags:
      - currency
      summary: Get currency rate by id
      operationId: get_currency_rate_id
      parameters:
      - name: rate_id
        in: path
        description: The rate_id thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyRate'
        400:
          description: Invalid currency supplied
        404:
          description: Currency not found
      x-openapi-router-controller: swagger_server.controllers.currency_controller
    put:
      tags:
      - currency
      summary: Updates currency rate
      description: This is only can be done by the logged in user
      operationId: update_currency_rate
      parameters:
      - name: rate_id
        in: path
        description: rate_id that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update currency rate object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrencyRate'
        required: true
      responses:
        400:
          description: Invalid rate supplied
        404:
          description: rate not found
      x-openapi-router-controller: swagger_server.controllers.currency_controller
    delete:
      tags:
      - currency
      summary: Delete currency rate
      description: This can only be done by the logged in user.
      operationId: delete_currency_rate
      parameters:
      - name: rate_id
        in: path
        description: The rate_id that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        400:
          description: Invalid rate supplied
        404:
          description: rate not found
      x-openapi-router-controller: swagger_server.controllers.currency_controller
  /posting:
    post:
      tags:
      - posting
      summary: Create posting
      operationId: create_posting
      requestBody:
        description: Created posting object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Posting'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.posting_controller
  /posting/{posting_id}:
    get:
      tags:
      - posting
      summary: Get posting by id
      operationId: get_posting_by_id
      parameters:
      - name: posting_id
        in: path
        description: The posting_id thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posting'
        400:
          description: Invalid posting supplied
        404:
          description: posting not found
      x-openapi-router-controller: swagger_server.controllers.posting_controller
    put:
      tags:
      - posting
      summary: Updates posting
      description: This is only can be done by the logged in user
      operationId: update_posting
      parameters:
      - name: posting_id
        in: path
        description: posting_id that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update posting object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Posting'
        required: true
      responses:
        400:
          description: Invalid posting supplied
        404:
          description: Posting not found
      x-openapi-router-controller: swagger_server.controllers.posting_controller
    delete:
      tags:
      - posting
      summary: Delete posting
      description: This can only be done by the logged in user.
      operationId: delete_posting
      parameters:
      - name: posting_id
        in: path
        description: The posting_id that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        400:
          description: Invalid posting supplied
        404:
          description: posting not found
      x-openapi-router-controller: swagger_server.controllers.posting_controller
  /posting/{posting_id}/publish:
    post:
      tags:
      - posting
      summary: Publish the posting
      operationId: publish_posting
      parameters:
      - name: posting_id
        in: path
        description: The posting_id thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Publish posting object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Posting'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.posting_controller
  /posting/{posting_id}/sell:
    post:
      tags:
      - posting
      summary: Sell posting
      operationId: sell_posting
      parameters:
      - name: posting_id
        in: path
        description: The posting_id thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: sell posting object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Posting'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.posting_controller
  /posting/gallery:
    post:
      tags:
      - posting
      summary: Create posting photo
      operationId: create_posting_photo
      requestBody:
        description: Created posting object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostingPhotoGallery'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.posting_controller
  /posting/gallery/{photo_id}:
    get:
      tags:
      - posting
      summary: Get posting photo by id
      operationId: get_posting_photo_by_id
      parameters:
      - name: photo_id
        in: path
        description: The photo thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostingPhotoGallery'
        400:
          description: Invalid currency supplied
        404:
          description: Currency not found
      x-openapi-router-controller: swagger_server.controllers.posting_controller
    put:
      tags:
      - posting
      summary: Updates posting photo
      description: This is only can be done by the logged in user
      operationId: update_posting_photo
      parameters:
      - name: photo_id
        in: path
        description: photo_id that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update photo rate object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostingPhotoGallery'
        required: true
      responses:
        400:
          description: Invalid photo supplied
        404:
          description: photo not found
      x-openapi-router-controller: swagger_server.controllers.posting_controller
    delete:
      tags:
      - posting
      summary: Delete posting photo
      description: This can only be done by the logged in user.
      operationId: delete_posting_photo
      parameters:
      - name: photo_id
        in: path
        description: The photo that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        400:
          description: Invalid photo supplied
        404:
          description: photo not found
      x-openapi-router-controller: swagger_server.controllers.posting_controller
...
components:
  schemas:
...
    Currency:
      required:
      - code
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        active:
          type: boolean
          default: true
      example:
        code: code
        active: true
        id: 0
    CurrencyRate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        price:
          type: number
        date:
          type: string
          format: date
        active:
          type: boolean
          default: true
        currency_id:
          type: integer
          format: int64
      example:
        date: 2000-01-23
        price: 6.027456183070403
        active: true
        id: 0
        currency_id: 1
...
    Posting:
      required:
      - category_id
      - currency_id
      - id
      - posting_type
      - sales_price
      - title
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        sales_price:
          type: number
        currency_id:
          type: integer
          format: int64
        posting_type:
          type: string
          default: fixed_sales_price
          enum:
          - fixed_sales_price
          - negotiable_sales_price
          - on_request_sales_price
        state:
          type: string
          default: draft
          enum:
          - draft
          - published
          - selled
        category_id:
          type: integer
          format: int64
        active:
          type: boolean
          default: true
      example:
        category_id: 5
        sales_price: 6.027456183070403
        posting_type: fixed_sales_price
        description: description
        active: true
        id: 0
        state: draft
        title: title
        currency_id: 1
    PostingPhotoGallery:
      required:
      - id
      - image_path
      - posting_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        posting_id:
          type: integer
          format: int64
        image_path:
          type: string
        active:
          type: boolean
          default: true
      example:
        image_path: image_path
        active: true
        posting_id: 6
        id: 0
...